1.3.03
Petite amélioration graphique des hitvalues, qui volent dans le World Space au lieu du Screen Space (ca rend pas mal)
Ptite update du menu screen pour que l'UI marche aussi là bas
Petit fix du fait que l'on pouvait dropper de l'armure que l'on avait équipé
Et le shop remarche aussi

+++ ProjetTuteuré/Assets/Scripts/ButtonInventaire.cs
@@ -79,7 +79,11 @@ public class ButtonInventaire : MonoBehaviour, IPointerClickHandler
             var objet = inv.listeItems[(int)char.GetNumericValue(gameObject.name[9])];
             if(objet != null)
             {
-                if(objet != player.GetComponent<Player>().armeCorpsACorpsEquipee && objet != player.GetComponent<Player>().armeDistanceEquipee)
+                if(objet != player.GetComponent<Player>().armeCorpsACorpsEquipee &&
+                    objet != player.GetComponent<Player>().armeDistanceEquipee &&
+                    objet != player.GetComponent<Player>().armureJambesEquipee &&
+                    objet != player.GetComponent<Player>().armureTeteEquipee &&
+                    objet != player.GetComponent<Player>().armureTorseEquipee)
                 {
                     objet.GetComponent<Item>().drop();
                 }
